requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
httpx==0.25.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
anyio==3.7.1
